Consolidate helpers folder into utils
Why:
We currently have two somewhat similar folders of `helpers` and `utils`,
we would like to have a single place for utility functions.
