+++ src/ExposureHistory/History.spec.tsx
@@ -3,7 +3,7 @@ import { fireEvent, wait, cleanup, render } from "@testing-library/react-native"
 import { useNavigation } from "@react-navigation/native"
 
 import { toExposureHistory } from "../gaen/exposureNotifications"
-import { DateTimeUtils } from "../helpers"
+import { DateTimeUtils } from "../utils"
 import { factories } from "../factories"
 
 import History from "./History"
