+++ src/gaen/exposureNotifications.spec.ts
@@ -1,6 +1,6 @@
 import dayjs from "dayjs"
 
-import { DateTimeUtils } from "../helpers"
+import { DateTimeUtils } from "../utils"
 import { Possible, ExposureDatum, ExposureInfo } from "../exposureHistory"
 import {
   toExposureInfo,
