--- src/helpers/index.ts
@@ -1,3 +1,0 @@
-import * as DateTimeUtils from "./dateTimeUtils"
-
-export { DateTimeUtils }
